feat: rename common/service to common/helper
