+++ src/common/helper/statisticsService/helpers.js
