+++ src/common/helper/statisticsService/statisticsService.js
