+++ src/component/common/auth-provider/service.js
@@ -12,14 +12,12 @@ export const fetchStatistics = async () => {
   return statistics;
 };
 
-export const fetchData = () => async (dispatch, getState) => {
-  const store = getState();
+export const fetchData = () => async (dispatch) => {
+  await dispatch(getUserSettings());
 
-  await getUserSettings(dispatch);
+  dispatch(fetchDictionaryWords());
 
-  await fetchDictionaryWords(dispatch, store);
-
-  await fetchUserWords(dispatch);
+  dispatch(fetchUserWords());
 
   // await fetchStatistics();
 };
